fix: circular dependencies
