+++ packages/compose-react/src/components/common/app-switcher/app-switcher.types.ts
@@ -1,0 +1,15 @@
+import type { ServiceName } from "@/store";
+import type { RuntimeKey } from "@/types";
+
+export interface BlocksApp {
+  key: ServiceName;
+  label: string;
+  description: string;
+  url: string;
+  icon: React.ReactNode;
+  clientId: RuntimeKey;
+  redirectUri: RuntimeKey;
+  initiateUrl: string;
+  isLoading: boolean;
+  isDisabled: boolean | (() => boolean);
+}
