+++ packages/compose-react/src/components/common/filter-toolbar/filter-controls.tsx
@@ -1,0 +1,19 @@
+import { DateRange } from "./date-range/date-range";
+import { DropdownSearchInput } from "./dropdown-search-input/dropdown-search-input";
+import { MultiSelect } from "./multi-select/multi-select";
+import { Radio } from "./radio/radio";
+import { ResetButton } from "./reset-button/reset-button";
+import { SearchInput } from "./search-input/search-input";
+import { SortHeader } from "./sort-header";
+import { TablePagination } from "./table-pagination";
+
+export const FilterControls = {
+  Radio,
+  MultiSelect,
+  SearchInput,
+  DropdownSearchInput,
+  DateRange,
+  ResetButton,
+  SortHeader,
+  TablePagination,
+};
