+++ packages/compose-react/src/components/common/filter-toolbar/index.ts
@@ -1,30 +1,17 @@
-import { DateRange } from "./date-range/date-range";
-import { DropdownSearchInput } from "./dropdown-search-input/dropdown-search-input";
-import { MultiSelect } from "./multi-select/multi-select";
-import { Radio } from "./radio/radio";
-import { ResetButton } from "./reset-button/reset-button";
-import { SearchInput } from "./search-input/search-input";
-import { SortHeader } from "./sort-header";
-import { TablePagination } from "./table-pagination";
+export {
+  FilterToolbar,
+  FilterToolBarMobileView,
+  type FilterChangeHandler,
+} from "./filter-toolbar";
 
-export const FilterControls = {
-  Radio,
-  MultiSelect,
-  SearchInput,
-  DropdownSearchInput,
-  DateRange,
-  ResetButton,
-  SortHeader,
-  TablePagination,
-};
-
-export * from "./filter-toolbar";
+export { FilterControls } from "./filter-controls";
 
 export {
   useSortQueryParams,
   sortQueryParsers,
   type SortValue,
 } from "./sort-header";
+
 export {
   usePaginationQueryParams,
   paginationQueryParsers,
