+++ packages/compose-react/src/pages/dashboard/dashboard-overview-sections/core-api/core-api-group-section.tsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
 import { ChevronDown } from "lucide-react";
-import { Badge } from "@/components";
+import { Badge } from "@/components/core/badge";
 import {
   Collapsible,
   CollapsibleContent,
   CollapsibleTrigger,
-} from "@/components";
+} from "@/components/core/collapsible";
 import { cn } from "@/lib/utils";
 import { CoreApiEndpointRow } from "./core-api-endpoint-row";
 import type { ICoreApiEndpoint } from "./core-api-endpoint.model";
