+++ packages/compose-react/src/pages/dashboard/dashboard-overview-sections/core-api/index.ts
@@ -1,2 +1,6 @@
-export * from "./core-api-card";
-export * from "./core-api-endpoint.model";
+export { CoreApiCard } from "./core-api-card";
+export type {
+  ICoreApiEndpoint,
+  HttpMethod,
+  ISwaggerDocument,
+} from "./core-api-endpoint.model";
