+++ packages/compose-react/src/pages/login/login.constant.ts
@@ -1,4 +1,4 @@
-import type { BlocksProduct } from "./blocks-login";
+import type { BlocksProduct } from "./login.types";
 
 export const BLOCKS_PRODUCTS: BlocksProduct[] = [
   {
